backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
pydantic==2.9.2
pydantic-settings==2.5.2
google-generativeai==0.8.2
python-dotenv==1.0.1

